Added Idea Chains and Testing Scripts
Added idea chains to replace links as links were essentially subpar ideas. Idea chains also allow direction when build ideas and give a path.

The testing directory contains a python script that will allow for easy writing of testing scripts (in python as well) that test features of the project. The testing script is a single point fo entry that runs all tests in the testing directory

Added a todo list with things to do